backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-dotenv==1.0.1
python-multipart>=0.0.9
PyJWT[crypto]>=2.8.0
psycopg2-binary>=2.9.0
boto3>=1.34.116
anthropic>=0.40.0
httpx
